Add archives support
This patch adds the ability to work with images inside archives
directly, without unpacking the whole archive on the disk. To work
with archives, the libarchive library is used.

By default, archives support is disabled because it brings a new build
and runtime dependency from libarchive, and it may break someone's
builds. Use 'archive=1' compile switch to enable archives support.

Imlib2 has its own limited support for compressed images, but it only
works with a single compressed image, and doesn't allow it to work
with multiple images inside one archive.

To reference a single image inside an archive from a file list, a new
URI scheme was introduced. Files from an archive are represented in
the following form: 'archive://path/to/archive.file:path/to/image'.

Creating a list of all files inside an archive may be a time-consuming
process. And by default, feh will treat archives like regular files
without trying to look for images inside of them. To actually force
feh to use archives as a directory with files, a new option
'--recursive-archives' should be used.